backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
mysql-connector-python
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
email-validator
pytest
httpx
pytest-asyncio
passlib[bcrypt] python-multipart